Moved subscription group screen to its own package
