+++ lib/home/_following.dart
@@ -3,9 +3,9 @@ import 'dart:developer';
 import 'package:dotted_border/dotted_border.dart';
 import 'package:flutter/material.dart';
 import 'package:fritter/database/entities.dart';
-import 'package:fritter/home/home.dart';
+import 'package:fritter/group/group_screen.dart';
+import 'package:fritter/home/home_screen.dart';
 import 'package:fritter/home_model.dart';
-import 'package:fritter/subscription_group.dart';
 import 'package:fritter/user.dart';
 import 'package:provider/provider.dart';
 import 'package:pull_to_refresh/pull_to_refresh.dart';
